Fixed Issue 83; Transforming graphics context was introducing slight rounding errors in the font metrics measurements. Calculating string widths using compnent.getGraphics() would yield slightly smaller widths than calculating them using the active Graphics object.
The fix is to calculate the "minimum acceptable" width that erring components (JButtons and JLabels) can be given the current graphics context and checking to see if it's larger than the minimum width the JLabel is reporting (which is computed at creation time assuming an identity transform). If it's larger (ellipsis would be shown) then it changes the minimum size of the component to be the the nearest ceil(minAcceptableWidth).

The calculation for the minimum size involves measuring text, adding insets, affording for icons and icon gaps.

Also, when calculating the bounds of the component, insets were not being considered but should have been.

Only hackish element of this fix is that its only applied to JButton and JLabels right now since I think they're the only components I can think of that calculate their minimum sizes in this way.

git-svn-id: http://piccolo2d.googlecode.com/svn/piccolo2d.java/trunk@550 aadc08cf-1350-0410-9b51-cf97fce99a1b